backend/requirements.txt
fastapi
uvicorn[standard]
sqlmodel
asyncpg
psycopg2-binary
pgvector
python-jose[cryptography]
passlib[bcrypt]
bcrypt<4.0.0
python-multipart
requests
redis
ldap3
google-auth
simple-salesforce
aiomysql
cryptography
opencv-python
numpy
Pillow
qrcode
openai
google-cloud-vision
boto3